ibmetrics: summary functions to metrics module
Start moving functions out of the report script and into modules.

The previous get_summary() function is split now: one function
calculates the values and returns a dictionary and another creates the
human-readable string that can be printed as a summary.

The new functions drop the calculation of the average packages per build
(including and excluding empty values).
We don't care about this right.  It can be added back when we need it.