metrics: optionally squash clouds for footprint metrics
When working with footprints, split or merge clouds using an option.
By default they're split, to maintain the old behaviour, but this might
change soon.

Minor bug fix: we accidentally mapped `aws` to `aws` instead of `ami` to
`aws`.  This meant that nothing mapped to `aws` since the image type is
called `ami`.  This had no effect on our metrics and figures since we
instead plotted `ami` in place of `aws`.